Refactor send attachments
